refactored email template sending
